useDummyAuth hook and moving entire dummyUI to dummyUIBtn
